% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{fst}
\alias{fst}
\title{Extract FST from a population-level kinship matrix or vector of inbreeding coefficients}
\usage{
fst(x, w)
}
\arguments{
\item{x}{The vector of inbreeding coefficients \eqn{(f_j^T)}, or the kinship matrix \eqn{\Phi^T} if \code{x} is a matrix.}

\item{w}{Weights for individuals (optional, defaults to uniform weights)}
}
\value{
\eqn{F_{ST}}{FST}
}
\description{
This function simply returns the weighted mean inbreeding coefficient \eqn{f_j^T}.
If no weights are provided, the regular mean \eqn{f_j^T} is returned.
If a kinship matrix \eqn{\Phi^T} is provided, then \eqn{f_j^T} are extracted from its diagonal using \code{\link{inbr}} (assumes the diagonal of \eqn{\Phi^T} is \eqn{\phi_{jj}^T = \frac{1}{2}(1+f_j^T)}{\phi_jj^T = (1+f_j^T)/2} as \code{\link{popkin}} returns, and not \eqn{f_j^T} as \code{\link{inbrDiag}} returns).
}
\details{
The returned weighted mean inbreeding coefficient equals the generalized \eqn{F_{ST}}{FST} if all individuals are "locally outbred" (i.e. if the self relatedness of every individual stems entirely from the population structure rather than due partly to having unusually closely related parents, such as first or second cousins).
Note most individuals in population-scale human data are locally outbred.
If there are locally inbred individuals, the returned value will overestimate \eqn{F_{ST}}{FST}.
}
\examples{
\dontrun{
## Get FST from a genotype matrix
## This example assumes input is in BED format and is loaded using BEDMatrix
## "file" is path to BED file (excluding .bed extension)
library(BEDMatrix)
X <- BEDMatrix(file) # load genotype matrix object
Phi <- popkin(X, subpops) # calculate kinship from genotypes and subpopulation labels
w <- weightsSubpops(subpops) # can weigh individuals so subpopulations are balanced
Fst <- fst(Phi, w) # use kinship matrix and weights to calculate fst

Fst <- fst(Phi) # no weights implies uniform weights

inbr <- inbr(Phi) # if you extracted inbr for some other analysis...
Fst <- fst(inbr, w) # ...use this inbreeding vector as input too!
}

}
