% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAEMinSubpops.R
\name{getAEminSubpops}
\alias{getAEminSubpops}
\title{Estimate the minimum expected value of A using subpopulations}
\usage{
getAEminSubpops(A, subpops)
}
\arguments{
\item{A}{The \eqn{n \times n} A matrix from \code{getA}.}

\item{subpops}{The length-\eqn{n} vector of subpopulation assignments for each individual.}
}
\value{
The estimate of the minimum expected value of A (scalar).
}
\description{
Estimate the minimum expected value of A using subpopulations
}
\examples{
\dontrun{AEMinHat <- getAEminSubpops(A, subpops)}

}
