% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFst.R
\name{getFst}
\alias{getFst}
\title{Extract \eqn{F_{ST}} from a kinship matrix or a vector of inbreeding coefficients}
\usage{
getFst(x, w)
}
\arguments{
\item{x}{The vector of inbreeding coefficients, or the kinship matrix (if class(x) is matrix).}

\item{w}{Weights for individuals to use in calculating \eqn{F_{ST}} (optional, defaults to uniform weights)}
}
\value{
\eqn{F_{ST}}, which is the weighted mean inbreeding coefficient.
}
\description{
Extract \eqn{F_{ST}} from a kinship matrix or a vector of inbreeding coefficients
}
\examples{
\dontrun{
Fst <- getFst(inbr, w) # use inbreeding vector as input
Fst <- getFst(Phi, w) # use kinship matrix as input
Fst <- getFst(Phi) # no weights implies uniform weights
}

}
