% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInbr.R
\name{getInbr}
\alias{getInbr}
\title{Extract inbreeding coefficients from a kinship matrix}
\usage{
getInbr(Phi)
}
\arguments{
\item{Phi}{The kinship matrix, with self-kinship values along the diagonal}
}
\value{
The length-\eqn{n} vector of inbreeding coefficients for each individual.
}
\description{
Extract inbreeding coefficients from a kinship matrix
}
\examples{
\dontrun{inbr <- getInbr(Phi)}

}
