% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKinshipFromA.R
\name{getKinshipFromA}
\alias{getKinshipFromA}
\title{Estimate kinship from A and an estimate of the minimum expected A}
\usage{
getKinshipFromA(A, AEMin)
}
\arguments{
\item{A}{The \eqn{n \times n} A matrix from \code{getA}.}

\item{AEMin}{The estimate of the minimum expected value of A.}
}
\value{
The estimated \eqn{n \times n} kinship matrix, with a minimum expected value of zero.
}
\description{
Estimate kinship from A and an estimate of the minimum expected A
}
\examples{
\dontrun{Phi <- getKinshipFromA(A, AEmin)}

}
