% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgdp_subset.R
\docType{data}
\name{hgdp_subset}
\alias{hgdp_subset}
\title{HGDP subset}
\format{
a matrix of 0's, 1's and 2's.
}
\source{
Stanford HGDP \url{http://www.hagsc.org/hgdp/files.html}
}
\usage{
hgdp_subset
}
\value{
genotype matrix
}
\description{
Subset of the HGDP dataset.
}
