% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbr.R
\name{inbr}
\alias{inbr}
\title{Extract inbreeding coefficients from a kinship matrix}
\usage{
inbr(Phi)
}
\arguments{
\item{Phi}{The \eqn{n \times n}{n-by-n} kinship matrix \eqn{\Phi^T}.}
}
\value{
The length-\eqn{n} vector of inbreeding coefficients \eqn{f_j^T} for each individual \eqn{j}.
}
\description{
The kinship matrix \eqn{\Phi^T} contains inbreeding coefficients \eqn{f_j^T} along the diagonal, present as \eqn{\phi_{jj}^T = \frac{1}{2}(1+f_j^T)}{\phi_jj^T = (1+f_j^T)/2}.  This function extracts the vector of \eqn{f_j^T} values from the input \eqn{\Phi^T}.
}
\examples{
\dontrun{inbr <- inbr(Phi)}

}
