% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbr_diag.R
\name{inbrDiag-deprecated}
\alias{inbrDiag-deprecated}
\title{Replace kinship diagonal with inbreeding coefficients}
\usage{
inbrDiag(kinship)
}
\arguments{
\item{kinship}{A kinship matrix}
}
\value{
The modified kinship matrix, with inbreeding coefficients along the diagonal.
}
\description{
Replace kinship diagonal with inbreeding coefficients
}
\seealso{
\code{\link{popkin-deprecated}}
}
\keyword{internal}
