% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minAvgSubpops.R
\name{minAvgSubpops}
\alias{minAvgSubpops}
\title{Estimate the minimum expected value of a matrix \eqn{Y} using subpopulations}
\usage{
minAvgSubpops(Y, subpops = NULL)
}
\arguments{
\item{Y}{A symmetric \eqn{n \times n} matrix with values between every individual pair, including self comparisons.}

\item{subpops}{A length-\eqn{n} vector of subpopulation assignments for each individual.  If missing, every individual is effectively treated as a different population.}
}
\value{
The scalar estimate of the minimum expected value of \eqn{Y}.
}
\description{
This function averages the values of a matrix \eqn{Y} between every subpopulation pair and returns the minimum of these averages.
The return value can be used to adjust an \eqn{A} matrix to yield the kinship matrix, or rescale a given kinship matrix (see below).
}
\details{
When \eqn{Y=A} is the \eqn{A} matrix, the output is a stable estimate of \eqn{\hat{A}_{Emin}}.
When \eqn{Y=\Phi} is a kinship matrix, the output is a stable estimate of the minimum kinship \eqn{\phi_{min}}.
If the kinship values between the most distant pair of subpopulations is truly zero, then the estimated \eqn{\hat{A}_{Emin}} can be used to adjust \eqn{A} into consistent kinship estimates using \code{\link{getKinshipFromA}}, or the estimated \eqn{\phi_{min}} can be used to rescale the pre-existing kinship matrix \eqn{\Phi} used as input using \code{\link{rescalePopkin}}.

If no subpopulation partition is provided, the function returns the minimum value of \eqn{Y}.
This default choice may be appropriate in some settings, but is susceptible to bias when there are few SNPs and many pairs of individuals with zero kinship (taking the most extreme estimate is clearly worse than averaging these values would be).
This default is provided for convenience, to explore the data when a correct choice of subpopulations is not clear, but is not recommended as a final approach.
}
\examples{
\dontrun{
## the recommended form using appropriate subpopulation labels
AEMinHat <- minAvgSubpops(A, subpops)
## a simple default for exploratory analysis
AEMinHat <- minAvgSubpops(A) # == min(A)
}
}
