% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_eff.R
\name{neff-deprecated}
\alias{neff-deprecated}
\title{Calculates the effective sample size of the data}
\usage{
neff(kinship, max = TRUE, w = NULL, retW = TRUE, nonneg = TRUE,
algo = c('gradient', 'newton', 'heuristic'), tol = 1e-10)
}
\arguments{
\item{kinship}{A kinship matrix}

\item{max}{Return maximum \eqn{n_{eff}}{n_eff}; otherwise compute \eqn{n_{eff}}{n_eff} using provided weights.}

\item{w}{Weights for individuals.}

\item{retW}{Return weights along with \eqn{n_{eff}}{n_eff}.}

\item{nonneg}{Optimize constrained to non-negative weights.}

\item{algo}{Algorithm for finding optimal non-negative weights: gradient, newton, or heuristic.}

\item{tol}{Tolerance parameter for "gradient" and "newton" algorithms.}
}
\value{
A list containing \eqn{n_{eff}}{n_eff} and the weights.
}
\description{
Calculates the effective sample size of the data
}
\seealso{
\code{\link{popkin-deprecated}}
}
\keyword{internal}
