% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_popkin.R
\name{plotPopkin-deprecated}
\alias{plotPopkin-deprecated}
\title{Visualize one or more kinship matrices}
\usage{
plotPopkin(
                      x,
                      colCont,
                      coln = 100,
                      xMar = NULL,
                      marPad = 0.2,
                      diagLine = FALSE,
                      panelLetters = toupper(letters),
                      panelLetterCex = 1.5,
                      ylabAdj = NA,
                      ylabLine = 0,
                      addLayout = TRUE,
                      nr = 1, 
                      legTitle = 'Kinship',
                      legMar = NULL,
                      nPretty = 5,
                      showNames = FALSE,
                      namesCex = 1,
                      namesLine = NA,
                      labsCex = 1,
                      labsLas = 0,
                      labsLine = 0,
                      labsSkipLines = FALSE,
                      labsLwd = 1,
                      labsCol = 'black',
                      labsDoTicks = FALSE,
                      labsDoText = TRUE,
                      labsEven = FALSE,
                      ...
)
}
\arguments{
\item{...}{Params to pass to new plot_popkin}
}
\value{
Nothing
}
\description{
Visualize one or more kinship matrices
}
\seealso{
\code{\link{popkin-deprecated}}
}
\keyword{internal}
