% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_popkin.R
\name{rescalePopkin-deprecated}
\alias{rescalePopkin-deprecated}
\title{Rescale kinship matrix to set a given kinship value to zero.}
\usage{
rescalePopkin(kinship, subpops = NULL, min_kinship = NA)
}
\arguments{
\item{kinship}{A kinship matrix}

\item{subpops}{Vector of subpopulation assignments for each individual.}

\item{min_kinship}{A scalar kinship value to define the new zero kinship.}
}
\value{
The rescaled kinship matrix, with the desired level of relatedness set to zero.
}
\description{
Rescale kinship matrix to set a given kinship value to zero.
}
\seealso{
\code{\link{popkin-deprecated}}
}
\keyword{internal}
