% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_subpops.R
\name{weightsSubpops-deprecated}
\alias{weightsSubpops-deprecated}
\title{Get weights for individuals that balance subpopulations}
\usage{
weightsSubpops(subpops)
}
\arguments{
\item{subpops}{The length-\eqn{n} vector of subpopulation assignments for each individual.}
}
\value{
The length-\eqn{n} vector of weights for each individual.
}
\description{
Get weights for individuals that balance subpopulations
}
\seealso{
\code{\link{popkin-deprecated}}
}
\keyword{internal}
