% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbrDiag.R
\name{inbrDiag}
\alias{inbrDiag}
\title{Replace kinship diagonal with inbreeding coefficients}
\usage{
inbrDiag(Phi)
}
\arguments{
\item{Phi}{The kinship matrix with self-kinship values along the diagonal}
}
\value{
The modified kinship matrix, with inbreeding coefficients along the diagonal
}
\description{
The usual kinship matrix contains self-kinship values \eqn{(1+f_j)/2} where \eqn{f_j} are inbreeding coefficients.
This function returns a modified kinship matrix with the diagonal replaced with the \eqn{f_j} values.
This form produces more aesthetically pleasing visualizations, but is not appropriate for modeling (e.g. in GWAS or heritability).
}
\examples{
\dontrun{PhiMod <- inbrDiag(Phi)}

}
