% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popkin-package.R
\docType{package}
\name{popkin-package}
\alias{popkin}
\alias{popkin-package}
\title{popkin: a package for estimating kinship and FST under arbitrary population structure}
\description{
The key function of this package is also named \code{\link{popkin}}, which estimates the kinship matrix of all individual pairs from their genotype matrix.
Inbreeding coefficients, the generalized \eqn{F_{ST}}{FST}, and the individual-level pairwise \eqn{F_{ST}}{FST} matrix are extracted from the kinship matrix using \code{\link{inbr}}, \code{\link{fst}}, and \code{\link{pwfst}}, respectively.
\code{\link{fst}} accepts weights for individuals to balance subpopulations obtained with \code{\link{weightsSubpops}}.
Kinship matrices can be renormalized (to change the most recent common ancestor population or MRCA) using \code{\link{rescalePopkin}}.
Lastly, kinship and pairwise \eqn{F_{ST}}{FST} matrices can be visualized using \code{\link{plotPopkin}} (with the help of \code{\link{inbrDiag}} for kinship matrices only).
}
\examples{
\dontrun{
## estimate and visualize kinship and FST from a genotype matrix

## This example assumes input is in BED format and is loaded using BEDMatrix
## "file" is path to BED file (excluding .bed extension)
library(BEDMatrix)
X <- BEDMatrix(file) # load genotype matrix object

## "subpops" below is a vector of subpopulation assignments for individuals

## estimate the kinship matrix "Phi" from the genotypes "X"!
## all downstream analysis require "Phi", none use "X" after this
Phi <- popkin(X, subpops) # calculate kinship from X and optional subpop labels

## plot the kinship matrix, marking the subpopulations
## note inbrDiag replaces the diagonal of Phi with inbreeding coefficients
plotPopkin( inbrDiag(Phi), labels=subpops )

## extract inbreeding coefficients from Phi
inbr <- inbr(Phi)

## estimate FST
w <- weightsSubpops(subpops) # weigh individuals so subpopulations are balanced
Fst <- fst(Phi, w) # use kinship matrix and weights to calculate fst
Fst <- fst(inbr, w) # estimate more directly from inbreeding vector (same result)

## estimate the pairwise FST matrix
pwF <- pwfst(Phi)
## ... and visualize it
legTitle <- expression(paste('Pairwise ', F[ST])) # fancy legend label
plotPopkin(pwF, labs=subpops, legTitle=legTitle) # NOTE no need for inbrDiag() here!

## rescale the kinship matrix using different subpopulations (implicitly changes the MRCA)
Phi2 <- rescalePopkin(Phi, subpops2)
}

}
\author{
\strong{Maintainer}: Alejandro Ochoa \email{ochoa@princeton.edu}

}
