% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwfst.R
\name{pwfst}
\alias{pwfst}
\title{Estimate the individual-level pairwise FST matrix}
\usage{
pwfst(Phi)
}
\arguments{
\item{Phi}{The \eqn{n \times n}{n-by-n} kinship matrix}
}
\value{
The \eqn{n \times n}{n-by-n} pairwise \eqn{F_{ST}}{FST} matrix
}
\description{
This function construct the individual-level pairwise \eqn{F_{ST}}{FST} matrix implied by the input kinship matrix.
If the input is the true kinship matrix, the return value corresponds to the true pairwise \eqn{F_{ST}}{FST} matrix.
On the other hand, if the input is the estimated kinship returned by \code{\link{popkin}}, then the return value is the pairwise \eqn{F_{ST}}{FST} estimates described in our paper.
In all cases the diagonal of the pairwise \eqn{F_{ST}}{FST} matrix is zero by definition.
}
\examples{
\dontrun{pwF <- pwfst(Phi)}
 
}
